make the fetch format validate paradigm generic
