--- internal/model/README.md
@@ -1,1 +1,0 @@
-This directory will hold all of the structs of the objects to be used when converting from API request, to populating go objects, which are easier to work with
