+++ server.go
@@ -11,7 +11,7 @@ import (
 	"github.com/99designs/gqlgen/graphql/handler/transport"
 	"github.com/99designs/gqlgen/graphql/playground"
 	"github.com/cpreciad/transit/graph"
-	"github.com/cpreciad/transit/internal"
+	apiqe "github.com/cpreciad/transit/internal/api_query_engine"
 	"github.com/vektah/gqlparser/v2/ast"
 )
 
@@ -26,7 +26,7 @@ func main() {
 	// this should be where a resolver type should be injected
 
 	srv := handler.New(graph.NewExecutableSchema(graph.Config{Resolvers: &graph.Resolver{
-		QueryEngine: internal.NewApiQueryEngine(),
+		QueryEngine: apiqe.NewApiQueryEngine(),
 	}}))
 
 	srv.AddTransport(transport.Options{})
